Add test coverage and remove condition from ShowBreadcrumb
This commit adds a spec for both the HasBreadcrumb and ShowBreadcrumb
modules.  In order to enable tests to run considerable support assets
have been added including, models, gem dependencies, and the Rails
framework.

The reason for the test support files is because this gem extends
ActiveRecord models via mixing these two modules.  This means in order
to test the expected behavior a full working ActiveRecord environment is
required.  This is achieved by including Rails and building up a partial
app via the /support directory.

In addition to the test coverage, a condition in the ShowBreadcrumb
module was found to not be required.  Therefore it was removed.

Lastly, the README was touched up to better word the functionality of
this gem.