Add gem dependencies and meta info to gemspec
This commit adds semantic versions to the rails-spec
and sqlite3 gems.  Also added was a home page pointed to the
gem source code hosted at Github.com.  Lastly, the description and
summary was improved to be more descriptive of the gem and what it does.

This commit was in response to warnings encountered during the gem build
process.