Rename has_breadcrumb references to simple_breadcrumbs
This commit changes references of has_breadcrumb to simple_breadcrumbs
at the upper 'gemspec' level.  Has_breadcrumb is left intact at the
library level since there is no risk of collision at that level.

The reason for this due to the name 'has_breadcrumb' already being taken
on rubygems.org.

[25539]