Update rspec to version 3.0
In order to keep the test suite up to date for simple_breadcrumbs, this
commit updates rspec to version 3.0. It removes the deprecated should
syntax and instead uses expect syntax. In addition, it removes the
treat_symbols_as_metadata_keys_with_true_values rspec config option as
this is now the default behavior.