Update dependencies for Rails 5.1 compatibility
This commit updates the gems dependencies to allow it to be used with
Rails 5.1. To accomplish this, it increases the acceptable version range
for Active Record and Active Support for Rails 5.1, and removes the
version lock for rspec-rails to make it possible to use the proper
supported version corresponding to Active Record and Active Support.
No changes to the code were required.

[612147225717591](https://app.asana.com/0/612147225717591/612147225717591)