Use correct yaml key to use the api key
In order to push up a gem to rubygems with the api key, this commit
sets the correct yaml key with the rubygems api key. This commit also
bumps the version to satisfy version constraints on the action. This
commit also removes the need for the `Bearer` in the api key environment
variable.